chore: move tab components into their own folder
